+++ src/gui/pluginspage.cpp
@@ -18,10 +18,8 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <QDockWidget>
 #include <QFileDialog>
 #include <QMessageBox>
-#include <QSettings>
 
 #include "ui_pluginspage.h"
 #include "pluginspage.h"
@@ -36,50 +34,85 @@ PluginsPage::PluginsPage(MainGUI* mainGUI)
     : QWidget(mainGUI)
     , m_ui(new Ui_PluginsPage)
     , m_mainApp(mainGUI->mainApp())
-    , m_innerWindow(new QMainWindow())
 {
     m_ui->setupUi(this);
-
-    m_innerWindow->setStyleSheet("QMainWindow { background-color: rgb(24,24,24); }");
-    m_innerWindow->setCentralWidget(m_ui->table);
-    m_ui->dockBrowser->setTitleBarWidget(nullptr);
-    m_innerWindow->addDockWidget(Qt::RightDockWidgetArea, m_ui->dockBrowser);
-    m_ui->iwLayout->addWidget(m_innerWindow);
 
     m_ui->labelPlugins->setFont(FontStyles::subtitle1());
 
     connect(m_ui->bImport, SIGNAL(pressed()), SLOT(importPlugin()));
-    connect(m_ui->table, SIGNAL(itemSelectionChanged()), SLOT(rowSelectionChanged()));
+    connect(m_ui->table, SIGNAL(itemSelectionChanged()),
+            SLOT(rowSelectionChanged()));
 
-    for (Plugin* p : m_mainApp->plugins()) { insertRow(p); }
+    for (Plugin* p : m_mainApp->plugins()) {
+        insertRow(p);
+    }
     connect(m_mainApp, SIGNAL(pluginAdded(const Plugin*)),
             SLOT(insertRow(const Plugin*)));
 
-    QSettings s;
-    m_innerWindow->restoreGeometry(s.value("gui/pluginsPageGeometry").toByteArray());
-    m_innerWindow->restoreState(s.value("gui/pluginsPageState").toByteArray());
+    m_ui->bUnload->setVisible(false);
+    m_ui->bReload->setVisible(false);
+    connect(m_ui->bUnload, SIGNAL(pressed()), SLOT(slotUnload()));
+    connect(m_ui->bReload, SIGNAL(pressed()), SLOT(slotReload()));
 }
 
 PluginsPage::~PluginsPage()
 {
-    QSettings s;
-    s.setValue("gui/pluginsPageGeometry", m_innerWindow->saveGeometry());
-    s.setValue("gui/pluginsPageState", m_innerWindow->saveState());
     delete m_ui;
+}
+
+void PluginsPage::slotUnload()
+{
+    auto key = m_ui->table->currentItem()->data(Qt::UserRole).value<PluginKey>();
+    Plugin* plugin = m_mainApp->plugins().value(key);
+    if (!plugin) {
+        return;
+    }
+
+    int res = QMessageBox::question(this, "Unloading Plugin",
+        "Are you sure you want to unload the plugin '"+plugin->id()+"'?");
+    if (res == QMessageBox::No) {
+        return;
+    }
+
+    QString error;
+    if (m_mainApp->unloadPlugin(plugin, error)) {
+        int row = m_ui->table->currentRow();
+        m_ui->table->clearSelection();
+        m_ui->table->removeRow(row);
+    } else {
+        QMessageBox::warning(this, "Unloading Plugin", error);
+    }
+}
+
+void PluginsPage::slotReload()
+{
+    auto key = m_ui->table->currentItem()->data(Qt::UserRole).value<PluginKey>();
+    Plugin* plugin = m_mainApp->plugins().value(key);
+    if (!plugin) {
+        return;
+    }
+
+    QString error;
+    if (m_mainApp->reloadPlugin(plugin, error)) {
+        int row = m_ui->table->currentRow();
+        m_ui->table->clearSelection();
+        m_ui->table->removeRow(row);
+    } else {
+        QMessageBox::warning(this, "Reloading Plugin", error);
+    }
 }
 
 void PluginsPage::rowSelectionChanged()
 {
-    int row = m_ui->table->currentRow();
-    if (row < 0) {
-        return;
-    }
+    bool rowIsSelected = !m_ui->table->selectedItems().empty();
+    m_ui->bUnload->setVisible(rowIsSelected);
+    m_ui->bReload->setVisible(rowIsSelected);
+    if (!rowIsSelected) { return; }
 
-    PluginKey k(m_ui->table->item(row, UID)->text(),
-                m_ui->table->item(row, VERSION)->text().toInt());
-    const Plugin* plugin = m_mainApp->graph(k);
+    auto key = m_ui->table->currentItem()->data(Qt::UserRole).value<PluginKey>();
+    const Plugin* plugin = m_mainApp->graph(key);
     if (!plugin) {
-        plugin = m_mainApp->model(k);
+        plugin = m_mainApp->model(key);
     }
 
     loadHtml(plugin);
@@ -130,58 +163,21 @@ void PluginsPage::insertRow(const Plugin* plugin)
     } else {
         qFatal("invalid plugin type! It should never happen.");
     }
-    typeItem->setData(Qt::UserRole, static_cast<int>(plugin->type()));
 
-    QPushButton* bUnload = new QPushButton();
-    bUnload->setFlat(true);
-    bUnload->installEventFilter(new ButtonHoverWatcher(this));
+    QTableWidgetItem* uidItem = new QTableWidgetItem(plugin->id());
+    auto keyv = QVariant::fromValue(plugin->key());
+    typeItem->setData(Qt::UserRole, keyv);
+    uidItem->setData(Qt::UserRole, keyv);
 
     m_ui->table->setSortingEnabled(false);
     int row = m_ui->table->rowCount();
     m_ui->table->insertRow(row);
-    m_ui->table->setItem(row, UID, new QTableWidgetItem(plugin->id()));
-    m_ui->table->setItem(row, VERSION, new QTableWidgetItem(QString::number(plugin->version())));
     m_ui->table->setItem(row, TYPE, typeItem);
-    m_ui->table->setCellWidget(row, UNLOAD, bUnload);
+    m_ui->table->setItem(row, UID, uidItem);
     m_ui->table->setSortingEnabled(true);
 
-    m_ui->table->item(row, VERSION)->setTextAlignment(Qt::AlignCenter);
-
     m_ui->table->horizontalHeader()->setSectionResizeMode(UID, QHeaderView::ResizeToContents);
-    m_ui->table->horizontalHeader()->setSectionResizeMode(VERSION, QHeaderView::ResizeToContents);
     m_ui->table->horizontalHeader()->setSectionResizeMode(TYPE, QHeaderView::ResizeToContents);
-    m_ui->table->horizontalHeader()->setSectionResizeMode(UNLOAD, QHeaderView::ResizeToContents);
-
-    connect(bUnload, &QPushButton::pressed, [this, typeItem, plugin]() {
-        int res = QMessageBox::question(this, "Unloading Plugin",
-            "Are you sure you want to unload the plugin '"+plugin->id()+"'?");
-        if (res == QMessageBox::No) {
-            return;
-        }
-
-        QString error;
-        if (m_mainApp->unloadPlugin(plugin, error)) {
-            m_ui->table->clearSelection();
-            m_ui->table->removeRow(typeItem->row());
-        } else {
-            QMessageBox::warning(this, "Unloading Plugin", error);
-        }
-    });
 }
 
-bool ButtonHoverWatcher::eventFilter(QObject* watched, QEvent* event)
-{
-    auto button = qobject_cast<QPushButton*>(watched);
-    if (!button) {
-        return false;
-    } else if (event->type() == QEvent::Enter) { // hovered
-        button->setIcon(QIcon(":/icons/x.png"));
-        return true;
-    } else if (event->type() == QEvent::Leave){
-        button->setIcon(QIcon());
-        return true;
-    }
-    return false;
-}
-
-}
+} // evoplex